GrossProfitModel/requirements.txt
azure-functions
pandas
numpy
scikit-learn
requests
azure-storage-blob
openpyxl
lxml
xgboost
azure-storage-file-share
html5lib
bs4
